src/test/java/dev/dobicinaitis/advent/Day5Test.java
package dev.dobicinaitis.advent;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day5Test {

    private final Puzzle puzzle = new Day5();

    @Test
    void firstPuzzleExampleTest() {
        assertEquals(35L, puzzle.getFirstPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    void firstPuzzleMyInputTest() {
        assertEquals(218513636L, puzzle.getFirstPuzzleSolution(puzzle.myInputFilename));
    }

    @Test
    void secondPuzzleExampleTest() {
        assertEquals(46L, puzzle.getSecondPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    void secondPuzzleMyInputTest() {
        assertEquals(81956384L, puzzle.getSecondPuzzleSolution(puzzle.myInputFilename));
    }
}